refactor: 設定管理ロジックを useSettings フックへ分離
- useSettings.ts: showSettings の管理と handleSaveSettings を移譲
- useSettings.test.ts: 設定管理ロジック単体のテストを追加
- useApp.ts: useSettings を使用するようにリファクタリング
- useApp.test.tsx: 統合テストとしての最小限の内容に整理
