+++ src/hooks/useApp.ts
@@ -4,23 +4,23 @@ import {
   useUpdateBookmark,
   useDeleteBookmark,
 } from './useBookmarks'
-import { getOrigin, validateApiUrl } from '@shared/utils/url'
 import {
-  LOG_MESSAGES,
   UI_MESSAGES,
-  COMMON_MESSAGES,
   HTML_ATTRIBUTES,
 } from '@shared/constants'
-import { useQueryClient } from '@tanstack/react-query'
-import { useApi } from '../contexts/ApiContext'
 import { useBookmarkReorder } from './useBookmarkReorder'
+import { useSettings } from './useSettings'
 import type { BookmarkId } from '@shared/schemas/bookmark'
 
 export const useApp = () => {
   const [selectedId, setSelectedId] = useState<BookmarkId | null>(null)
-  const [showSettings, setShowSettings] = useState(false)
-  const queryClient = useQueryClient()
-  const { apiUrl: currentApiUrl, updateApiUrl } = useApi()
+  const {
+    showSettings,
+    currentApiUrl,
+    toggleSettings,
+    closeSettings,
+    handleSaveSettings,
+  } = useSettings()
 
   const { data, isLoading, error } = useBookmarks()
   const updateMutation = useUpdateBookmark()
@@ -75,35 +75,6 @@ export const useApp = () => {
   const handleClose = useCallback(() => {
     setSelectedId(null)
   }, [])
-
-  const toggleSettings = useCallback(() => {
-    setShowSettings((prev) => !prev)
-  }, [])
-
-  const closeSettings = useCallback(() => {
-    setShowSettings(false)
-  }, [])
-
-  const handleSaveSettings = useCallback(
-    (newUrl: string): string | null => {
-      try {
-        const error = validateApiUrl(newUrl)
-        if (error) return error
-
-        const sanitizedUrl = getOrigin(newUrl)
-        updateApiUrl(sanitizedUrl)
-        queryClient.clear()
-        setShowSettings(false)
-        return null
-      } catch (err) {
-        console.error(LOG_MESSAGES.EXTENSION_SETTING_SAVE_FAILED, err)
-        return err instanceof Error
-          ? err.message
-          : COMMON_MESSAGES.UNKNOWN_ERROR
-      }
-    },
-    [updateApiUrl, queryClient],
-  )
 
   return {
     bookmarks,
